test(webview): move the suites into their own tests/ folder
They sat next to the modules they cover, which meant the test glob, the
lint/format globs and tsconfig.test.json each had to name both core/ and ui/.
One folder is one entry everywhere, and adding a suite for a third directory
no longer means touching four patterns.

Imports gain a ../core/ hop; nothing else changes.
