+++ src/Corsinvest.VisualStudio.Agents/Chat/WebViewSrc/package.json
@@ -10,11 +10,11 @@
         "build:dev": "node tools/gen-bridge.mjs && node esbuild.config.mjs --dev",
         "dev": "node tools/gen-bridge.mjs && node esbuild.config.mjs --watch",
         "typecheck": "tsc --noEmit && tsc --noEmit -p tsconfig.test.json",
-        "test": "node --test \"core/**/*.test.ts\" \"ui/**/*.test.ts\"",
-        "lint": "eslint \"{core,ui}/**/*.ts\"",
-        "lint:fix": "eslint --fix \"{core,ui}/**/*.ts\"",
-        "format": "prettier --write \"{core,ui}/**/*.ts\" \"*.ts\"",
-        "format:check": "prettier --check \"{core,ui}/**/*.ts\" \"*.ts\"",
+        "test": "node --test \"tests/**/*.test.ts\"",
+        "lint": "eslint \"{core,ui,tests}/**/*.ts\"",
+        "lint:fix": "eslint --fix \"{core,ui,tests}/**/*.ts\"",
+        "format": "prettier --write \"{core,ui,tests}/**/*.ts\" \"*.ts\"",
+        "format:check": "prettier --check \"{core,ui,tests}/**/*.ts\" \"*.ts\"",
         "check": "npm run typecheck && npm run lint && npm run format:check && npm run test"
     },
     "dependencies": {
